Add workflows to default page props Get workflows from the project with getDefaultPageProps, so that project workflows are also available on the Classify page.
