+++ packages/app-project/pages/projects/[owner]/[project]/index.js
@@ -1,2 +1,8 @@
-export { default, getServerSideProps } from '@screens/ProjectHomePage'
+import getDefaultPageProps from '@helpers/getDefaultPageProps'
+export { default } from '@screens/ProjectHomePage'
+
+export async function getServerSideProps({ params, query, req, res }) {
+  const { props } = await getDefaultPageProps({ params, query, req, res })
+  return ({ props })
+}
