+++ packages/app-project/src/screens/ProjectHomePage/index.js
@@ -1,3 +1,2 @@
 export { default } from './ProjectHomePageContainer'
-export { getServerSideProps } from './getServerSideProps'
